requirements.txt
# Core Data Processing
apache-airflow==2.7.3
pyspark==3.5.0
kafka-python==2.0.2
confluent-kafka==2.3.0

# AWS Services
boto3==1.34.0
awscli==1.32.0

# Database and Storage
snowflake-connector-python==3.6.0
google-cloud-bigquery==3.14.0
psycopg2-binary==2.9.9
pymongo==4.6.1

# API Integration
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
plaid-python==16.0.0

# ML and Data Science
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
mlflow==2.10.0
joblib==1.3.2

# Visualization
plotly==5.18.0
dash==2.14.2
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
python-json-logger==2.0.7
schedule==1.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.13

# Development
black==23.12.1
flake8==7.0.0
pylint==3.0.3
